requirements.txt
flask
flask-cors
python-dotenv
requests
transformers
llama-cpp-python
sentence-transformers
faiss-cpu
PyMuPDF
etuptools 
wheel
ntlk
python -c "import nltk; nltk.download('punkt')"
python -c "import nltk; nltk.download('punkt_tab')"